fix(phase-5a): T2 — query key namespace + sale_count int cast
Two code-quality findings from review:

1. src/hooks/queries/keys.ts — rename inventoryConsumption and
   inventoryVariance key roots from "reports" (which collides
   with the Cash Close `reports(businessDate)` factory) to
   "inventory-reports". Prevents accidental cache blast from
   future broad invalidation of ["reports"].

2. database/002_functions.sql — cast count(distinct
   source_order_id) to ::int instead of ::bigint, and declare
   sale_count int in the return type. supabase-js returns
   Postgres bigint as a JS string (overflow safety), which
   would have broken the TS `sale_count: number` typing
   downstream. int is safe at coffee-shop scale (2^31 = 2B).

   database/tests/160_inventory_reports.sql Test 4 updated from
   2::bigint to 2::int. DROP + CREATE used in 002_functions.sql
   to allow OUT-column type change.

verify:phase: 75 Vitest + 99 pgTAP = 174 green.

+++ database/002_functions.sql
@@ -3084,7 +3084,7 @@ create or replace function public.inventory_consumption_by_ingredient(
   ingredient_name  text,
   unit             text,
   total_consumed   numeric,
-  sale_count       bigint
+  sale_count       int
 )
 language sql
 stable
@@ -3094,8 +3094,8 @@ as $$
     i.id           as ingredient_id,
     i.name as ingredient_name,
     i.unit,
-    sum(abs(sm.quantity_delta))::numeric        as total_consumed,
-    count(distinct sm.source_order_id)::bigint  as sale_count
+    sum(abs(sm.quantity_delta))::numeric      as total_consumed,
+    count(distinct sm.source_order_id)::int   as sale_count
   from public.stock_movements sm
   join public.ingredients i on i.id = sm.ingredient_id
   where sm.reason = 'sale_theoretical'
